定期爬蟲專用/Spider_man/requirements.txt
appnope==0.1.2
APScheduler==3.7.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
future==0.18.2
greenlet==1.0.0
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
line-bot-sdk==1.19.0
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.7
tabulate==0.8.9
tornado==6.1
traitlets==4.3.3
twstock==1.3.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
